unified folding logic
